// Client/src/store/actionTypes.ts
export enum ActionTypes {
  LOAD_USER_REQUEST = 'LOAD_USER_REQUEST',
  LOAD_USER_SUCCESS = 'LOAD_USER_SUCCESS',
  LOAD_USER_ERROR = 'LOAD_USER_ERROR',
  LOAD_USERS_LOCAL_STORAGE = 'LOAD_USERS_LOCAL_STORAGE',
  LOAD_USER_RESET = 'LOAD_USER_RESET',

  ADD_PRODUCT_TO_CART_REQUEST = 'ADD_PRODUCT_TO_CART_REQUEST',
  ADD_PRODUCT_TO_CART_SUCCESS = 'ADD_PRODUCT_TO_CART_SUCCESS',
  ADD_PRODUCT_TO_CART_ERROR = 'ADD_PRODUCT_TO_CART_ERROR',

  REMOVE_PRODUCT_FROM_CART_RQUEST = 'REMOVE_PRODUCT_FROM_CART_RQUEST',
  REMOVE_PRODUCT_FROM_CART_SUCCESS = 'REMOVE_PRODUCT_FROM_CART_SUCCESS',
  REMOVE_PRODUCT_FROM_CART_ERROR = 'REMOVE_PRODUCT_FROM_CART_ERROR',

  MINUS_PRODUCT_IN_CART_REQUEST = 'MINUS_PRODUCT_IN_CART_REQUEST',
  MINUS_PRODUCT_IN_CART_SUCCESS = 'MINUS_PRODUCT_IN_CART_SUCCESS',
  MINUS_PRODUCT_IN_CART_ERROR = 'MINUS_PRODUCT_IN_CART_ERROR',

  PLUS_PRODUCT_IN_CART_REQUEST = 'PLUS_PRODUCT_IN_CART_REQUEST',
  PLUS_PRODUCT_IN_CART_SUCCESS = 'PLUS_PRODUCT_IN_CART_SUCCESS',
  PLUS_PRODUCT_IN_CART_ERROR = 'PLUS_PRODUCT_IN_CART_ERROR',

  PLACE_PURCHASE_ORDER_REQUEST = 'PLACE_PURCHASE_ORDER_REQUEST',
  PLACE_PURCHASE_ORDER_SUCCESS = 'PLACE_PURCHASE_ORDER_SUCCESS',
  PLACE_PURCHASE_ORDER_ERROR = 'PLACE_PURCHASE_ORDER_ERROR',
  PLACE_PURCHASE_ORDER_NOT_ASKED = 'PLACE_PURCHASE_ORDER_NOT_ASKED',

  CLEAR_PRODUCT_IN_CART_REQUEST = 'CLEAR_PRODUCT_IN_CART_REQUEST',
  CLEAR_PRODUCT_IN_CART_SUCCESS = 'CLEAR_PRODUCT_IN_CART_SUCCESS',
  CLEAR_PRODUCT_IN_CART_ERROR = 'CLEAR_PRODUCT_IN_CART_ERROR',

  REMOVE_RENT_ORDER_REQUEST = 'REMOVE_RENT_ORDER_REQUEST',
  REMOVE_RENT_ORDER_SUCCESS = 'REMOVE_RENT_ORDER_SUCCESS',
  REMOVE_RENT_ORDER_ERROR = 'REMOVE_RENT_ORDER_ERROR',

  LOGIN_USER_REQUEST = 'LOGIN_USER_REQUEST',
  LOGIN_USER_SUCCESS = 'LOGIN_USER_SUCCESS',
  LOGIN_USER_ERROR = 'LOGIN_USER_ERROR',

  ADD_INVENTORY_REQUEST = 'ADD_INVENTORY_REQUEST',
  ADD_INVENTORY_SUCCESS = 'ADD_INVENTORY_SUCCESS',
  ADD_INVENTORY_ERROR = 'ADD_INVENTORY_ERROR',
  
  REMOVE_INVENTORY_REQUEST = 'REMOVE_INVENTORY_REQUEST',
  REMOVE_INVENTORY_SUCCESS = 'REMOVE_INVENTORY_SUCCESS',
  REMOVE_INVENTORY_ERROR = 'REMOVE_INVENTORY_ERROR',
  
  REMOVE_PRODUCT_REQUEST = 'REMOVE_PRODUCT_REQUEST',
  REMOVE_PRODUCT_SUCCESS = 'REMOVE_PRODUCT_SUCCESS',
  REMOVE_PRODUCT_ERROR = 'REMOVE_PRODUCT_ERROR',

  RENT_PRODUCT_REQUEST = 'RENT_PRODUCT_REQUEST',
  RENT_PRODUCT_SUCCESS = 'RENT_PRODUCT_SUCCESS',
  RENT_PRODUCT_ERROR = 'RENT_PRODUCT_ERROR',

  SIGN_UP_USER_REQUEST = 'SIGN_UP_USER_REQUEST',
  SIGN_UP_USER_SUCCESS = 'SIGN_UP_USER_SUCCESS',
  SIGN_UP_USER_ERROR = 'SIGN_UP_USER_ERROR',
  SIGN_UP_USER_RESET = 'SIGN_UP_USER_RESET',

  EDIT_USER_REQUEST = 'EDIT_USER_REQUEST',
  EDIT_USER_SUCCESS = 'EDIT_USER_SUCCESS',
  EDIT_USER_ERROR = 'EDIT_USER_ERROR',

  ADD_ADMIN_REQUEST = 'ADD_ADMIN_REQUEST',
  ADD_ADMIN_SUCCESS = 'ADD_ADMIN_SUCCESS',
  ADD_ADMIN_ERROR = 'ADD_ADMIN_ERROR',
  ADD_ADMIN_RESET = 'ADD_ADMIN_RESET',

  ADD_PRODUCT_REQUEST = 'ADD_PRODUCT_REQUEST',
  ADD_PRODUCT_SUCCESS = 'ADD_PRODUCT_SUCCESS',
  ADD_PRODUCT_ERROR = 'ADD_PRODUCT_ERROR',
  ADD_PRODUCT_RESET = 'ADD_PRODUCT_RESET',

  LOG_OUT_REQUEST = 'LOG_OUT_REQUEST',
  
  LOAD_ALL_PRODUCTS_REQUEST = 'LOAD_ALL_PRODUCTS_REQUEST',
  LOAD_ALL_PRODUCTS_SUCCESS = 'LOAD_ALL_PRODUCTS_SUCCESS',
  LOAD_ALL_PRODUCTS_ERROR = 'LOAD_ALL_PRODUCTS_ERROR',

  LOAD_ALL_ORDERS_REQUEST = 'LOAD_ALL_ORDERS_REQUEST',
  LOAD_ALL_ORDERS_SUCCESS = 'LOAD_ALL_ORDERS_SUCCESS',
  LOAD_ALL_ORDERS_ERROR = 'LOAD_ALL_ORDERS_ERROR',

  PLACE_RENT_ORDERS_REQUEST = 'PLACE_RENT_ORDERS_REQUEST',
  PLACE_RENT_ORDERS_SUCCESS = 'PLACE_RENT_ORDERS_SUCCESS',
  PLACE_RENT_ORDERS_ERROR = 'PLACE_RENT_ORDERS_ERROR',
  PLACE_RENT_ORDERS_NOT_ASKED = 'PLACE_RENT_ORDERS_NOT_ASKED',

  INVENTORY_RESET = 'INVENTORY_RESET',
};